requirements.txt
alembic==1.8.1
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.17.1
packaging==23.0
passlib==1.7.4
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2023.2
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
Werkzeug==2.2.2
zipp==3.15.0
